Seperated vosk from the main file, put it into speech_detection.py instead. Still not tested.
